+++ resources/store.py
@@ -1,13 +1,14 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.store import StoreModel
 
 class Store(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument(name="storename", required=True, help="a store name is required to proceed", type=str,case_sensitive=False)
-    parser.add_argument(name="user_id", required=True, help="only active users can create a store", type=str,case_sensitive=False)
+    parser.add_argument(name="user_id", required=True, help="only active users can create a store", type=int,case_sensitive=False)
     parser.add_argument(name="country", required=False, help="Country where the store is located", type=str,case_sensitive=False)
 
+    @jwt_required
     def get(self, storeid):
         store = StoreModel.find_by_id(storeid)
         if store:
@@ -15,8 +16,12 @@ class Store(Resource):
         else:
             return {"message" : "store not found"}, 404
 
+    @jwt_required
     def post(self):
+        claim = get_jwt_claims()
         data = Store.parser.parse_args()
+        if not claim["is_admin"] and claim["userid"] != data["user_id"]: 
+                return {"message" : "Admin priviledge required."}, 401
 
         if StoreModel.find_by_name(storename=data["storename"]):
             return {"message" : f"A store with name {data['storename']} already exists"}, 400
@@ -31,15 +36,21 @@ class Store(Resource):
         return store.json(), 201
 
     #use for authentication before calling post
+    @jwt_required
     def put(self, storeid):
+        claim = get_jwt_claims()
         data = Store.parser.parse_args()
+
+        if not claim["is_admin"] and claim["userid"] != data["user_id"]: 
+                return {"message" : "Admin priviledge required."}, 401
+
         store_byid = StoreModel.find_by_id(storeid=storeid)
         store_byname = StoreModel.find_by_name(storename=data["storename"])
 
         #check if store name is in use
         if store_byid:
             if store_byname:
-                if store_byname.storename!= store_byid.storename: return {"message" : f"store name {store_byname.storename} already exists."},400 # 400 is for bad request
+                if store_byname.storename != store_byid.storename: return {"message" : f"store name {store_byname.storename} already exists."},400 # 400 is for bad request
         
             #update
             try:
@@ -55,22 +66,29 @@ class Store(Resource):
                     store_byid = StoreModel(**data)
                     store_byid.save_to_db()
             except Exception as e:
-                print(f"error is {e} -- yy")
+                print(f"error is {e}")
                 return {"message" : "An error occured creating the item"}, 500 #Internal server error
     
         return store_byid.json(), 201
 
+    @jwt_required
     def delete(self, storeid):
+        claim = get_jwt_claims()
         store = StoreModel.find_by_id(storeid)
 
         if store:
+            if not claim["is_admin"] and claim["userid"] != store.user_id: 
+                return {"message" : "Admin priviledge required."}, 401
             store.delete_from_db()
             return {"message" : "Store deleted"}
+        elif not store and not claim["is_admin"]:
+            return {"message" : "Admin priviledge required."}, 401
 
         return {"message" : "Store not found"}
 
 
 class StoreList(Resource):
+    @jwt_required
     def get(self):
         stores = StoreModel.find_all()
         return {"stores" : [store.json() for store in stores]}
