+++ resources/user.py
@@ -1,13 +1,27 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (create_access_token, create_refresh_token, jwt_required, get_jwt_claims)
 from models.user import UserModel
 
-#class to list all user
-class UserList(Resource):
-    def get(self):        
-        users = UserModel.find_all()
-        if users: return {"users" : [ user.json() for user in users]},201
-        return {"message" : 'Item not found'}, 400
+#class to login users
+class UserLogin(Resource):
+    parser = reqparse.RequestParser()
+    parser.add_argument(name="password",   type=str,   required=True, help="password cannot be blank")
+    parser.add_argument(name="email",      type=str,   required=True, help="email cannot be blank")
+
+    @classmethod
+    def post(cls):
+        data = cls.parser.parse_args() # get data <1>
+        user = UserModel.find_by_email(data.get("email"))  # find user by email <2>
+
+        if user and user.password == data.get("password"): # check password <3>
+            access_token = create_access_token(identity=user.id, fresh=True) # create access token <4>
+            refresh_token = create_refresh_token(identity=user.id) # create refresh token <5>
+            return {
+                    "access_token" : access_token,
+                    "refresh_token" : refresh_token
+                    }, 200
+
+        return {"message" : "Invalid Credentials"}, 401
 
 
 #class to register user
@@ -44,6 +58,19 @@ class UserRegister(Resource):
         return user.json(), 201
 
 
+# class to list all user
+class UserList(Resource):
+    @jwt_required
+    def get(self):     
+        claim = get_jwt_claims()
+        if not claim["is_admin"]: 
+            return {"message" : "Admin priviledge required."}, 401
+
+        users = UserModel.find_all()
+        if users: return {"users" : [ user.json() for user in users]},201
+        return {"message" : 'Item not found'}, 400
+
+
 #class to create user and get user
 class User(Resource):
     parser = reqparse.RequestParser()
@@ -60,7 +87,12 @@ class User(Resource):
     parser.add_argument(name="state",      type=str,   required=False, help="user's state",case_sensitive=False) 
 
     @classmethod
-    def get(userid=None):
+    @jwt_required
+    def get(cls,userid=None):
+        claim = get_jwt_claims()
+        if not claim["is_admin"] and claim["userid"] != userid: 
+            return {"message" : "Admin priviledge required."}, 401
+
         user = UserModel.find_by_id(id=userid)
 
         if user: return {"user" : user.json()},201
@@ -68,8 +100,12 @@ class User(Resource):
 
     #use for authentication before calling post
     @classmethod
-    def put(userid):
-        
+    @jwt_required
+    def put(cls,userid):
+        claim = get_jwt_claims()
+        if not claim["is_admin"] and claim["userid"] != userid: 
+            return {"message" : "Admin priviledge required."}, 401
+
         data = User.parser.parse_args()
         user = UserModel.find_by_id(id=userid)
         email = UserModel.find_by_email(email=data["email"])
@@ -101,7 +137,12 @@ class User(Resource):
 
     #use for authentication before calling post
     @classmethod
-    def delete(userid):
+    @jwt_required
+    def delete(cls,userid):
+        claim = get_jwt_claims()
+        if not claim["is_admin"] and claim["userid"] != userid: 
+            return {"message" : "Admin priviledge required."}, 401
+            
         user = UserModel.find_by_id(id=userid,)
         if user:
             user.delete_from_db()
